Explicitly set ps columns to 300 to fix gdb execution
When running the app under the debugger, the renderer killer was not
working, because the `ps` command was not printing out the full command
line of the PS entry. `--type=renderer` was not found because it was cut
off.

This explicitly sets the number of columns to 300, which allows the
renderer type on the command line to show up.

Additionally, print out the number of renderers killed even when it is
zero, so that we can detect when no renderers are detected.